requirements.txt
anthropic==0.42.0
beautifulsoup4==4.12.2
flask==3.0.2
python-dotenv==1.0.1 